requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
minio==7.1.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
pydantic==1.8.2
pytest==6.2.5
passlib[bcrypt]==1.7.4
python-multipart